Move genre state down and create FormStep type
